requirements.txt
python-telegram-bot
google-genai
openai
aiohttp